+++ src/Commands/ChurnCommand.php
@@ -5,8 +5,8 @@ namespace Churn\Commands;
 use Churn\Factories\ResultsRendererFactory;
 use Churn\Logic\ResultsLogic;
 use Churn\Managers\FileManager;
-use Churn\Processes\ProcessFactory;
-use Churn\Processes\ProcessHandlerFactory;
+use Churn\Process\ProcessFactory;
+use Churn\Process\ProcessHandlerFactory;
 use function count;
 use function file_get_contents;
 use Symfony\Component\Console\Command\Command;
