+++ src/Process/ChurnProcess.php
@@ -1,7 +1,6 @@
 <?php declare(strict_types = 1);
 
-
-namespace Churn\Processes;
+namespace Churn\Process;
 
 use Churn\Values\File;
 use Symfony\Component\Process\Exception\ProcessFailedException;
