+++ tests/Unit/Process/Handler/SequentialProcessHandlerTest.php
@@ -1,11 +1,11 @@
 <?php declare(strict_types = 1);
 
-namespace Churn\Tests\Unit\Processes\Handler;
+namespace Churn\Tests\Unit\Process\Handler;
 
 use Churn\Collections\FileCollection;
-use Churn\Processes\ChurnProcess;
-use Churn\Processes\Handler\SequentialProcessHandler;
-use Churn\Processes\ProcessFactory;
+use Churn\Process\ChurnProcess;
+use Churn\Process\Handler\SequentialProcessHandler;
+use Churn\Process\ProcessFactory;
 use Churn\Tests\BaseTestCase;
 use Churn\Values\File;
 use Illuminate\Support\Collection;
